Implemented Checkout logic and photo service

+++ API/Models/DTO/Cart/CartRequests/CartHeaderRequest.cs
@@ -3,7 +3,7 @@
     public class CartHeaderRequest
     {
         public int Id { get; set; }
-        public long Total { get; set; }
+        public decimal Total { get; set; }
         public string CouponCode { get; set; }
     }
 }
