+++ API/Models/DTO/Coupon/CouponDto.cs
@@ -1,4 +1,4 @@
 ﻿namespace API.Models.DTO.Coupon
 {
-    public record CouponDto(string CouponCode, long DiscountAmount, long MinPrice);
+    public record CouponDto(string CouponCode, decimal DiscountAmount, decimal MinPrice);
 }
